refactor: move csv lockfile parsing into its own file
